chore(core,proxy,runtime): move test modules to #[path] siblings for ratchet headroom
Eight files this arc must grow sit at their frozen ratchet baselines and
cannot gain a line. Move each one's #[cfg(test)] module to a *_tests.rs
sibling via #[path], the pattern truncation.rs, kv_memory.rs, and models.rs
already set. No behaviour change: every test moved intact, none altered,
none added, none removed. The only non-move edits are the #[path] mod
declarations themselves and one unused import dropped from the parse half
of the props split, which clippy -D warnings would reject in its new home.

props.rs splits two ways by responsibility — fetch/parse tests
(props_parse_tests.rs, 61 lines) and the check_baseline comparison
(props_baseline_tests.rs, 404) — because the two halves share nothing but
fixtures. The baseline half stays over the 300 budget and is recorded, the
same standing the other over-budget siblings have.

Ratchet arithmetic, all recorded via --update in this diff: inference.rs
2553 -> 1544, sampling.rs 1747 -> 565, settings.rs 1012 -> 652,
dashboard.rs 1111 -> 553, props.rs 995 -> 570. Three files fall under the
budget entirely and leave the baseline: sampling_provenance.rs (was 377),
apply.rs (was 343), server_config.rs (was 493). Net entries 176 -> 173.

rustfmt reflowed four call sites across two siblings after the one-level
dedent (token-identical, whitespace only); everything else is byte-exact.
model_commands.rs and shared_args.rs carry no test modules, so the later
PRs that grow them take deliberate, visible baseline bumps instead.

Full pre-commit green.
